refactor: Qr 관련 DTO를 reservation → qr 패키지로 이동
