+++ src/main/java/org/example/popspace/dto/qr/QrReservationDTO.java
@@ -1,4 +1,4 @@
-package org.example.popspace.dto.reservation;
+package org.example.popspace.dto.qr;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
